Create a very simple script for Zola to show him how fast a computer can count to 1 million. WOW!
Spending time with Zola and immersing him gently to the world of coding was great. He was so impressed that a computer can count to 1 million and soon after he was running the script from the CLI. He was also quite interested to learn about RegEx and happy to see his name matched by the regex. Such is like, simple through the eyes of children. Innocent, empty mind, empty thought.